Changed and moved files for better modularity
