+++ js/api/post/request-call/delete-post-data.mjs
