--- js/api/posts/index.mjs
@@ -1,5 +1,0 @@
-import { getRequest } from "./request-type/get-request.mjs";
-
-export function postRequests() {
-   getRequest();
-}
