+++ js/handlers/posts/create-post-handler.mjs
@@ -1,4 +1,4 @@
-import { createRequest } from "../../api/posts/request-type/create-request.mjs";
+import { createRequest } from "../../api/post/request-type/create-request.mjs";
 import { validateInputs } from "../../form-validation/validatedInput.mjs";
 
 export function createPostHandler() {
