+++ js/handlers/posts/post-dropdown-handler.mjs
@@ -1,5 +1,5 @@
-import { deleteRequest } from "../../api/posts/request-type/delete-request.mjs";
-import { updateRequest } from "../../api/posts/request-type/update-request.mjs";
+import { deleteRequest } from "../../api/post/request-type/delete-request.mjs";
+import { updateRequest } from "../../api/post/request-type/update-request.mjs";
 import { closeEditPostHandler } from "./close-edit-post-handler.mjs";
 
 export function postDropdownHandler(responseOk) {
