+++ js/html-data/post/html-post-content.mjs
@@ -1,5 +1,5 @@
-import { displayPosts } from "../../../display-data/display-posts.mjs";
-import { displaySinglePostById } from "../../../display-data/display-single-post-by-id.mjs";
+import { displayPosts } from "../../display-data/display-posts.mjs";
+import { displaySinglePostById } from "../../display-data/display-single-post-by-id.mjs";
 import { createParseDoc } from "../createParseDoc.mjs";
 
 export function changeToHtmlPost(postVariables) {
