+++ js/index.mjs
@@ -1,26 +1,25 @@
 import { validatedForms } from "./form-validation/index.mjs";
 import { offlineFeedback } from "./display-data/display-logout-feedback.mjs";
 import { findPostsHandlers, postHandlers, signupHandlers } from "./handlers/index.mjs";
-import { postRequests } from "./api/posts/index.mjs";
 import { displayProfileData } from "./display-data/display-profile-data.mjs";
 import { displaySinglePostById } from "./display-data/display-single-post-by-id.mjs";
+import { getRequest } from "./api/post/request-type/get-request.mjs";
 
 const path = location.pathname;
 
 validatedForms();
-
 if (path === "/profile/view/index.html" || path === "/profile/view/") {
    postHandlers();
-   postRequests();
+   getRequest();
    displayProfileData();
 } else if (path === "/profile/home-feed/index.html" || path === "/profile/home-feed/") {
    postHandlers();
-   postRequests();
+   getRequest();
    findPostsHandlers();
-} else if (path === "/profile/post/index.html" || "/profile/post/" || "/profile/post/?id=109") {
+} else if (path === "/profile/post/index.html" || path === "/profile/post/") {
    displaySinglePostById();
-   postRequests();
-} else {
+   getRequest();
+} else if (path === "/profile/register/index.html" || path === "/profile/register/") {
    offlineFeedback();
    signupHandlers();
 }
